// src/types/profile.ts
export interface User {
  avatar: string;
  fullName: string;
  email: string;
  phone: string;
  dob: string;
  id: string;
}
export interface Business {
  id: string;
  name: string;
  address: string;
  state: string;
  country: string;
  zipCode: string;
  businessType: string;
  industry: string;
  currency: string;
  taxId: string;
  logo: string;
  banner: string;
  email: string;
  phone: string;
}